requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
langchain
langchain-community
pypdf
chromadb
sentence-transformers
faiss-cpu
langchain_huggingface
langchain_community
langchain 
langchain_google_genai
dotenv